moved classes to own namespace to avoid large index file
